requirements.txt
# Email and environment configuration
sendgrid>=6.9.7
python-dotenv>=0.19.0

# Web framework and API
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5  # for form data processing

# Database
pyodbc>=4.0.32  # for SQL Server connection
sqlalchemy>=1.4.23

# Security and validation
python-jose[cryptography]>=3.3.0  # for JWT handling
passlib[bcrypt]>=1.7.4  # for password hashing
pydantic>=1.8.2  # for data validation

# Testing
pytest>=6.2.5
pytest-asyncio>=0.15.1
httpx>=0.18.2  # for async HTTP requests in tests 